condensed notes to one page
