+++ src/client/features/plants/Notes.jsx
@@ -1,15 +1,99 @@
 import { useGetNotesQuery } from "../../store/Notes";
 import { useSelector } from "react-redux";
-import { selectToken } from "../auth/authSlice";
-import Note from "./Note";
+import { selectToken, useMeQuery } from "../auth/authSlice";
+
+//  function for rendering our createNoteForm  // window to pop up on screen
+function NoteCard({ note }) {
+
+    // const [editNote] = useEditNoteMutation();
+    // const [deleteNote] = useDeleteNoteMutation();
+
+//     const [note, setNote]= useState(each.note);
+
+    // // save edited note
+    // const save = async (evt) => {
+    //     evt.preventDefault();
+    //     editNote({ ...note, note});
+    // }
+
+    // // delete note
+    // const onDelete = async (evt) => {
+    //     evt.preventDefault();
+    //     deleteNote(note.id);
+    // };
+
+    return (
+        <li>
+            <h4>{note.id}</h4>
+            <h4>{note.note}</h4>
+             {/* <form onSubmit={save}>
+                 <input type="text" value={note} 
+                 onChange={(e) => setNote(e.target.value)} 
+                 required />
+                 <button>Save</button>
+                 <button onClick={onDelete}> 🞪 </button>
+             </form> */}
+        </li>
+    )
+}
+
+function NewNote() {
+    const [note, setNote] = useState("");
+
+    //     use the hooks to create the mutation functions we will use
+    const [createNote] = useCreateNoteMutation();
+
+//     const closePopupWindow = () => {
+//         "x" button for closing window
+//         /onclick close window
+//         //error if text!=== 0 {return "Are you sure you want to close without saving note?"}
+//     }
+//     const createNote = () => {
+//         "Save Note" button to add 
+//         /onclick save note and close window
+//         //error if text === 0 {return "There is no note to save"}
+//     }
+
+//     return (
+//         <>
+//             <form>
+//                 <textarea name="note" rows="10" cols="20" />
+//                 <button onClick={closePopupWindow} id="delete">X</button>
+//                 <button onClick={saveCreatedNote} id="submit">Create Note</button>
+//             </form>
+//         </>
+//     )
+// }
+    const create = async (evt) => {
+        evt.preventDefault();
+        createNote({ note });
+    };
+
+    return(
+        <form onSubmit={create}>
+            <input
+            type="text"
+            value={note}
+            onChange={(e) => setNote(e.target.value)}
+            required
+            />
+            <button>Create Note</button>
+        </form>
+    );
+}
 
 
 export default function Notes(){
     const token = useSelector(selectToken);
     const { data: notes, isLoading } = useGetNotesQuery();
     
-    console.log(notes);
+    // console.log(notes);
+
+    // function Note() {
+    //     notes.map(note)
+    // };
     
+
     if (!token) {
         return <p>You must be logged in to see your tasks.</p>;
     };
@@ -18,15 +102,48 @@ export default function Notes(){
         return (
         <div className="notes">
             <h2>My Plant Notes</h2>
-            <NewNote />
+            {/* <NewNote /> */}
             {isLoading && <p>Loading notes...</p>}
             {notes && (
-                <ul>
-                    {notes.map((note) => (
-                       <Note key={note.id} note={note} />
-                       ))}
+                <ul className="notes-list">
+                    {notes?.map((note) => (
+                        <NoteCard key={note.id} note={note} /> 
+                     ))} 
                 </ul>
             )}
         </div>
         );
     };
+
+    //~ function for rendering our editNoteForm  // window to pop up on screen
+// function editNoteForm() {
+
+//     const closePopupWindow = () => {  // "x" button for closing window
+//         ///onclick close window
+//         ////error if no change in text {return "No changes are saved"}
+//     }
+    //const saveChangedNote = () => {
+        // input field for text
+        // text.value = text[selected note id] ; This will be the previous saved Note
+        // "Edit Note" button to add 
+        ///onclick update state, delete previous data and close popup window
+    //}
+
+//     return (
+//         <>
+//             <form>
+//                 <textarea name="note" rows="10" cols="20" />
+//                 <button onClick={closePopupWindow} id="delete">X</button>
+//                 <button onClick={saveChangedNote} id="submit">Create Note</button>
+//             </form>
+//         </>
+//     )
+// };
+
+
+//getElement by id createNoteBTN
+//onclick createNoteBTN(createNoteForm)
+//getElement by id editBTN
+//onclick editBTN(editNoteForm)
+
+// }
