+++ src/server/api/auth/index.js
@@ -72,6 +72,8 @@ router.post("/login", async (req, res, next) => {
   }
 });
 
+
+// THIS IS A BUG! WILL CONTINUOUSLY RUN AND CRASH THE SYSTEM
 router.get("/", async (req, res, next) => {
   try {
     const id = +req.headers.authorization;
